+++ app/sentry.client.config.js
@@ -11,7 +11,7 @@ if (process.env.NODE_ENV !== "development") {
     dsn: SENTRY_DSN,
     environment: SENTRY_ENV,
     release: `visualization-tool@${BUILD_VERSION}`,
-    tracesSampleRate: 1.0,
+    tracesSampleRate: 0.1,
     ignoreErrors: [
       // The ResizeObserver error is actually not problematic
       // @see https://forum.sentry.io/t/resizeobserver-loop-limit-exceeded/8402