+++ app/sentry.edge.config.js
@@ -11,6 +11,6 @@ if (process.env.NODE_ENV !== "development") {
     dsn: SENTRY_DSN,
     environment: SENTRY_ENV,
     release: `visualization-tool@${BUILD_VERSION}`,
-    tracesSampleRate: 1.0,
+    tracesSampleRate: 0.1,
   });
 }